Move loading to store and move Controls out of App
